Fix bug in applyTags when modifyPackage changes "marked" property
In some cases, when tagged builds are dependencies of each other,
modifyPackage may change the "marked" property from what the user had
requested in the TagList, and these would subsequently get skipped when
it was their turn to be processed. This commit stores a copy of the user
selections and sets the "marked" property back to the user selection
during each build's turn, ensuring each gets processed properly.